feat!: refactor config file for note and journal path
instead of having one section of the config file for both note and
journal with only one field "folder_path", add that directly into the
general section of the config with "note/journal_folder_path".

BREAKING CHANGE: The previous config file with the journal and note
sections must be change to this new format.